requirements.txt
streamlit==1.31.0
langchain-groq==0.1.3
pydantic==2.6.1
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.1.12